tmm-macro-terminal/requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
curl_cffi==0.15.0
Flask==3.1.3
flask-cors==6.0.2
frozendict==2.4.7
gunicorn==25.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multitasking==0.0.12
numpy==2.4.4
packaging==26.0
pandas==3.0.2
peewee==4.0.4
platformdirs==4.9.6
protobuf==7.34.1
pycparser==3.0
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
rich==15.0.0
schedule==1.2.2
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.8
yfinance==1.2.1
